+++ Products/Reportek/RemoteFMEConversionApplication.py
@@ -184,8 +184,7 @@ class RemoteFMEConversionApplication(SimpleItem):
         self.retryFrequency = REQUEST.form.get("retryFrequency")
         self.app_name = REQUEST.form.get("app_name")
         self.nRetries = int(REQUEST.form.get("nRetries"))
-        if REQUEST is not None:
-            return self.manage_settings_html(manage_tabs_message="Saved changes.")
+        return self.manage_settings_html(manage_tabs_message="Saved changes.")
 
     def get_fme_token(self):
         """Retrieves the token from FME"""
@@ -214,7 +213,7 @@ class RemoteFMEConversionApplication(SimpleItem):
                     """ retrieve token: {}-{}""".format(resp.status_code, resp.content)
                 )
         except Exception as e:
-            logger.error(
+            logger.exception(
                 """FME authentication request failed. Could not"""
                 """ retrieve token: {}""".format(str(e))
             )
@@ -268,15 +267,14 @@ class RemoteFMEConversionApplication(SimpleItem):
                             ):
                                 latest[grp_prefix] = e_file.bobobase_modification_time()
             if not latest:
-                err = "No convertible files found in the envelope. "
-                "Convertible file extensions for this workflow: {}.".format(
-                    ", ".join(ext)
+                raise ValueError(
+                    "No convertible files found in the envelope. "
+                    "Convertible file extensions for this workflow: {}.".format(
+                        ", ".join(ext)
+                    )
                 )
-                raise ValueError(err)
             up_group = list(latest.keys())[
-                list(latest.values()).index(
-                    sorted(list(latest.values()), reverse=True)[0]
-                )
+                list(latest.values()).index(max(latest.values()))
             ]
             files = [
                 f
@@ -465,7 +463,7 @@ class RemoteFMEConversionApplication(SimpleItem):
             return upload_storage["paths"][-1]
         if shapefile and upload_storage["paths"]:
             for p in reversed(upload_storage["paths"]):
-                if p.endswith(".shp") or p.endswith(".zip"):
+                if p.endswith((".shp", ".zip")):
                     gmls = [
                         fid.split(".")[0]
                         for fid in env.objectIds("Report Document")