+++ Products/Reportek/updates/u20161026_add_missing_postingdate_to_feedbacks.py
@@ -93,7 +93,7 @@ def rebuild_tempc(app, tempc):
         try:
             fb = brain.getObject()
         except Exception as e:
-            logger.error(
+            logger.exception(
                 "Unable to retrieve object: {} due to {}".format(brain.getURL(), str(e))
             )
         if fb:
@@ -119,7 +119,7 @@ def add_missing_postingdate(app):
         try:
             obj = brain.getObject()
         except Exception as e:
-            logger.error(
+            logger.exception(
                 "Unable to retrieve object: {} due to {}".format(brain.getURL(), str(e))
             )
         if obj: