+++ Products/Reportek/views/list_users.py
@@ -180,7 +180,7 @@ class ListUsers(BaseAdmin):
 
         if path:
             parts = path.split("/")
-            if path.startswith("http") or path.startswith("/"):
+            if path.startswith(("http", "/")):
                 use_path = path
                 if path.startswith("http"):
                     use_path = "/{0}".format("/".join(parts[3:]))
@@ -226,7 +226,7 @@ class ListUsers(BaseAdmin):
                         )
                 if REPORTEK_DEPLOYMENT == DEPLOYMENT_BDR:
                     # Hide our internal user agent from search results
-                    if "bdr_folder_agent" in list(users.keys()):
+                    if "bdr_folder_agent" in users:
                         del users["bdr_folder_agent"]
 
                 if not users and role != "Reporter (Owner)":